+++ package.json
@@ -31,7 +31,7 @@
         "babel-preset-stage-0": "^6.24.1",
         "clean-webpack-plugin": "3.0.0",
         "css-loader": "3.6.0",
-        "eslint": "^7.4.0",
+        "eslint": "^7.9.0",
         "eslint-config-airbnb": "^18.2.0",
         "eslint-config-airbnb-base": "^14.2.0",
         "eslint-plugin-html": "6.0.2",
