+++ src/blocks/article/article.css
@@ -41,7 +41,7 @@
     padding: 40px 0 64px 80px;
     text-align: left;
 }
-/* 
-.search-results_hidden {
-  display: none;
-} */
+
+.article__link {
+  text-decoration: none;
+}
