+++ src/blocks/header/__container/header__container.css
@@ -11,13 +11,19 @@
   }
 }
 
+@media screen and (max-width: 715px) {
+  .header__container {
+    margin: 0 40px;
+  }
+}
+
 @media screen and (max-width: 635px) {
   .header__container {
     margin: 0 30px;
   }
 }
 
-@media screen and (max-width: 445px) {
+@media screen and (max-width: 460px) {
   .header__container {
     display: none;
   }
