+++ src/blocks/header/__logo/header__logo.css
@@ -10,9 +10,7 @@
   color: #1A1B22;
 }
 
-
-
-@media screen and (max-width: 445px) {
+@media screen and (max-width: 460px) {
   .header__logo_top {
     margin-left: 20px;
     margin-top: 0px;
