+++ src/blocks/menu/__button/_auth/menu__button_auth.css
@@ -5,7 +5,7 @@
   border-radius: 100px;
 }
 
-@media screen and (max-width: 445px) {
+@media screen and (max-width: 460px) {
   .menu__button_auth {
     border: none;
   }
