+++ src/blocks/menu/__button/_exit/menu__button_exit.css
@@ -22,3 +22,7 @@
 .menu__button_exit-pic_black {
   filter: invert(1);
 }
+
+.menu__button_exit-mobile {
+  margin: 10px 50px 0 100px;
+}
