+++ src/blocks/menu/__button/_saved-articles/menu__button_saved-articles.css
@@ -1,5 +1,5 @@
 .menu__button_saved-articles {
-  margin-left: 34px;
+  /* margin-left: 34px; */
   margin-right: 32px;
 }
