+++ src/blocks/menu/menu.css
@@ -5,7 +5,7 @@
   display: flex;
   flex-direction: row;
   list-style: none;
-  justify-content: space-around;
+  justify-content: center;
   width: auto;
 }
 
@@ -22,9 +22,9 @@
   }
 }
 
-@media screen and (max-width: 445px) {
+/* @media screen and (max-width: 460px) {
   .menu {
     padding-left: 0px
 
   }
-}
+} */
