+++ src/index.html
@@ -33,9 +33,6 @@
 
       <header class="mobile-menu"></header>
 
-
-
-
       <section class="loading loading_hidden">
         <h2 class="search-results search-results_hidden">Результаты поиска</h2>
         <div class="article__list"></div>
@@ -52,7 +49,7 @@
         <button class="loading__button loading__button_hidden">Показать ещё</button>
 
       </section>
-      <!-- <div class="article__list"></div> -->
+      
       <section class="author">
         <div class="author__container">
           <img src="<%=require('./images/author.png').default %>" alt="Аватар" class="author__pic">
