+++ src/js/api/NewsApi.js
@@ -9,7 +9,6 @@ export default class NewsApi {
   }
 
   getArticles(keyWord) {
-    // console.log('getArticles', `${this.url}?q=${keyWord}&pageSize=${this.pageSize}`);
     return fetch(`${this.url}/everything?q=${keyWord}&pageSize=${this.pageSize}&apiKey=${this.apiKey}`, {
       method: 'GET',
       headers: {
@@ -22,13 +21,11 @@ export default class NewsApi {
       });
   }
 
-  // eslint-disable-next-line class-methods-use-this
   parseResponce(res) {
     if (res.ok) {
       return res.json();
     }
 
-    // eslint-disable-next-line prefer-promise-reject-errors
     return Promise.reject(`Что-то пошло не так: ${res.status}`);
   }
 }
