+++ src/js/components/MobileMenu.js
@@ -4,44 +4,44 @@ const createMobileMenu = (userData, loggedinHeader = false) => {
   
     if (userData) {
       template = `
-      <input id="menu__toggle" type="checkbox" />
-      <p class="header__logo header__logo_top">NewsExplorer</p>
-      <label class="menu__btn" for="menu__toggle">
-        <span></span>
-      </label>
-      <ul class="menu__box">
-      <li class="mobile-menu__button"><a class="menu__item" href="index.html" class="mobile-menu__link">Главная</a>
-      </li>
-      <li class="mobile-menu__button"><a class="menu__item" href="loggedin.html"
-      class="mobile-menu__link">Сохранённые
-      статьи</a></li>
-      <li class="menu menu__button menu__button_exit popup__button_hidden"><a href="#" class="link menu__link_exit">${userData.userName}</a>
-      <img src="./images/Union.png" alt="дверь на выход" class="menu__button_exit-pic">
-    </li>
-              </ul>
+        <input id="menu__toggle" type="checkbox">
+        <p class="header__logo header__logo_top">NewsExplorer</p>
+        <label class="menu__btn ${loggedinHeader ? 'menu__btn_black' : ''}" for="menu__toggle">
+          <span></span>
+        </label>
+        <ul class="menu__box">
+          <li class="mobile-menu__button"><a class="menu__item" href="index.html" class="mobile-menu__link">Главная</a>
+            </li>
+            <li class="mobile-menu__button"><a class="menu__item" href="loggedin.html"
+            class="mobile-menu__link">Сохранённые
+            статьи</a></li>
+            <li class="menu menu__button menu__button_exit menu__button_exit-mobile"><a href="#" class="link menu__link_exit">${userData.userName}</a>
+            <img src="./images/Union.png" alt="дверь на выход" class="menu__button_exit-pic">
+          </li>
+        </ul>
             `;
     } else {
       template = `
-      <input id="menu__toggle" type="checkbox" />
+      <input id="menu__toggle" type="checkbox">
       <p class="header__logo header__logo_top">NewsExplorer</p>
       <label class="menu__btn" for="menu__toggle">
         <span></span>
       </label>
       <ul class="menu__box">
-      <li class="mobile-menu__button"><a class="menu__item" href="index.html" class="mobile-menu__link">Главная</a>
-  </li>
-  <li class="mobile-menu__button "><a href="#"
-  class="menu__item mobile-menu__link mobile-menu__link_auth">Авторизоваться</a></li>
-
+        <li class="mobile-menu__button"><a class="menu__item" href="index.html" class="mobile-menu__link">Главная</a>
+        </li>
+        <li class="mobile-menu__button "><a href="#"
+        class="menu__item mobile-menu__link">Авторизоваться</a></li>
+      </ul>
     `;
     }
-    rootNode.insertAdjacentHTML('beforeend', template);
+  rootNode.insertAdjacentHTML('beforeend', template);
+  
   };
   
  
-  const mobileAuth = document.querySelector('.mobile-menu__link_auth');
-  mobileAuth.addEventListener('click', popup.open);
   
+
 export default createMobileMenu;
