+++ src/loggedin.html
@@ -14,8 +14,8 @@
 
 <body>
   <main class="page">
-      <div class="header__container header__black"></div>
-      <header class="mobile-menu"></header>
+    <div class="header__container header__black"></div>
+    <header class="mobile-menu"></header>
 
     <section class="top"></section>
     <div class="loading__search">
